% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper.R
\name{build_vis_network}
\alias{build_vis_network}
\title{Build visualization network}
\usage{
build_vis_network(pkg)
}
\arguments{
\item{pkg}{A character string specifying the package name.}
}
\value{
A visNetwork object representing the visualization network.
}
\description{
This function builds a visualization network for a specified package.
}
