% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper.R
\name{filter_func_objs}
\alias{filter_func_objs}
\title{Filter function objects}
\usage{
filter_func_objs(pkg_exports)
}
\arguments{
\item{pkg_exports}{A tibble of package exports, as returned by get_package_contents().}
}
\value{
A filtered tibble of package exports.
}
\description{
This function filters the exported objects of a package to include only functions and S3/S4 methods/generics.
}
