% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper.R
\name{get_all_function_deps}
\alias{get_all_function_deps}
\title{Get all function dependencies}
\usage{
get_all_function_deps(pkg)
}
\arguments{
\item{pkg}{A character string specifying the package name.}
}
\value{
A tibble of function dependencies within the specified package.
}
\description{
This function retrieves all function dependencies within a specified package.
}
