% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper.R
\name{get_all_r6_method_deps}
\alias{get_all_r6_method_deps}
\title{Get all R6 method dependencies}
\usage{
get_all_r6_method_deps(pkg)
}
\arguments{
\item{pkg}{A character string specifying the package name.}
}
\value{
A tibble of R6 method dependencies within the specified package.
}
\description{
This function retrieves all R6 method dependencies within a specified package.
}
