% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper.R
\name{get_function_calls}
\alias{get_function_calls}
\title{Get function calls}
\usage{
get_function_calls(pkg, func)
}
\arguments{
\item{pkg}{A character string specifying the package name.}

\item{func}{A character string specifying the function name.}
}
\value{
A tibble of function calls within the specified function.
}
\description{
This function retrieves the function calls within a specified function in a package.
}
