% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper.R
\name{get_package_contents}
\alias{get_package_contents}
\title{Get package contents}
\usage{
get_package_contents(pkg_names)
}
\arguments{
\item{pkg_names}{A character vector of package names.}
}
\value{
A tibble with columns for the package name, function name, and function type.
}
\description{
This function retrieves the contents of a package, including the names and types of exported objects.
}
