% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper.R
\name{get_r6_method_calls}
\alias{get_r6_method_calls}
\title{Get R6 method calls}
\usage{
get_r6_method_calls(pkg, r6_class, r6_method, method_type = "public")
}
\arguments{
\item{pkg}{A character string specifying the package name.}

\item{r6_class}{A character string specifying the R6 class name.}

\item{r6_method}{A character string specifying the R6 method name.}

\item{method_type}{A character string specifying the method type (public or private).}
}
\value{
A tibble of method calls within the specified R6 method.
}
\description{
This function retrieves the method calls within a specified R6 method in a package.
}
