% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper.R
\name{get_r6_methods}
\alias{get_r6_methods}
\title{Get R6 methods}
\usage{
get_r6_methods(pkg)
}
\arguments{
\item{pkg}{A character string specifying the package name.}
}
\value{
A tibble of R6 methods within the specified package.
}
\description{
This function retrieves the methods of R6 classes within a specified package.
}
